requirements.txt
aiohttp>=3.8.0
beautifulsoup4>=4.12.2
yt-dlp>=2023.10.13
python-telegram-bot>=20.0
httpx>=0.23.0
gallery-dl>=1.25.0
asyncio>=3.4.3  # Встроен в Python, минимальная версия
lxml>=4.9.1  # Зависимость BeautifulSoup для парсинга
requests>=2.28.0  # Может использоваться yt-dlp или gallery-dl косвенно
psutil>=5.9.0  # Для мониторинга ресурсов (опционально, но полезно)
pyOpenSSL>=23.0.0  
certifi>=2023.7.22  